[feat] 홈화면 우체통 클릭시 데이터 조회 성공
