+++ app/src/main/java/com/toyou/toyouandroid/data/create/dto/response/HomeDto.kt
@@ -1,10 +1,10 @@
-package com.toyou.toyouandroid.data.home.dto.response
+package com.toyou.toyouandroid.data.create.dto.response
 
 import com.google.gson.annotations.SerializedName
 
 data class HomeDto(
-    @SerializedName("emotion") val emotion : String,
+    @SerializedName("emotion") val emotion : String?,
     @SerializedName("questionNum") val question : Int,
-    @SerializedName("cardId") val id : Int,
+    @SerializedName("cardId") val id : Int?,
     @SerializedName("uncheckedAlarm") val alarm : Boolean,
 )
