+++ app/src/main/java/com/toyou/toyouandroid/data/create/service/CreateService.kt
@@ -3,6 +3,7 @@ package com.toyou.toyouandroid.data.create.service
 import com.toyou.toyouandroid.data.create.dto.request.AnswerDto
 import com.toyou.toyouandroid.data.create.dto.response.AnswerPost
 import com.toyou.toyouandroid.data.create.dto.response.QuestionsDto
+import com.toyou.toyouandroid.data.create.dto.response.HomeDto
 import com.toyou.toyouandroid.network.BaseResponse
 import retrofit2.http.Body
 import retrofit2.http.GET
@@ -30,4 +31,9 @@ interface CreateService {
         @Path("cardId") card : Int,
         @Body request : AnswerDto
         ) : BaseResponse<Unit>
+
+    @GET("/users/home")
+    suspend fun getHomeEntry(
+        @Header("userId") id : Int,
+    ) : BaseResponse<HomeDto>
 }
